requirements.txt
streamlit
streamlit-mic-recorder
azure-ai-projects
azure-cosmos
azure-core
azure-identity
speechrecognition
pydub